chore(Queries): split qt monaco and editor wrap [YTFRONT-4802]
