+++ packages/ui/src/ui/pages/query-tracker/QueryEditor/QueryEditorMonaco.scss
@@ -1,0 +1,18 @@
+.yq-query-editor-monaco {
+    flex: 1;
+    border-bottom: 1px solid var(--dark-divider);
+    width: 100%;
+    height: calc(100% - 40px);
+    overflow: visible;
+
+    .selected-text {
+        z-index: 2;
+        position: absolute;
+        pointer-events: none;
+        opacity: 0.8;
+    }
+
+    .view-lines {
+        z-index: 3;
+    }
+}
